collect coupon

+++ src/main/java/cn/phpst/mall/repository/BannerRepository.java
@@ -1,12 +1,12 @@
 package cn.phpst.mall.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import cn.phpst.mall.model.Banner;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BannerRepository extends JpaRepository<Banner,Integer> {
+public interface BannerRepository extends JpaRepository<Banner, Long> {
     Banner findOneById(Integer id);
+
     Banner findOneByName(String name);
 }
