+++ src/main/java/cn/phpst/mall/repository/GridCategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GridCategoryRepository extends JpaRepository<GridCategory, Integer> {
+public interface GridCategoryRepository extends JpaRepository<GridCategory, Long> {
 }
