+++ src/main/java/cn/phpst/mall/repository/SpuRepository.java
@@ -7,9 +7,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpuRepository extends JpaRepository<Spu,Integer>  {
+public interface SpuRepository extends JpaRepository<Spu, Long> {
     Spu findOneById(Integer id);
 
     Page<Spu> findByCategoryIdOrderByCreateTimeDesc(Long cid, Pageable pageable);
+
     Page<Spu> findByRootCategoryIdOrderByCreateTime(Long cid, Pageable pageable);
 }
